LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-duplicate-names-list-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


FAIL TrustedTypePolicyFactory and policy list in CSP. Can't find variable: trustedTypes
